ApacheAirflow/pyproject.toml
[project]
name = "apacheairflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow[postgres]>=3.0.6",
]
